Moves ModelParams class into own file.
